Added a new function to apply to *edges* so that we get what we need
It would have been more natural for me to start from here, instead of making
first describe-paths and then this one. But is good for the purposes of the
book.
Funny way of writing a "map".

1]> (describe-paths 'living-room *edges*)
(THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE.)