Added new function to describe-objects at a given location.
More understandable after a while looking a lisp. But still a bit ugly ... why
mapcar and not just "map", why labels instead of anonymous or lambda functions,
why #', not so nice.
Anyways here is the result

[1]> (describe-objects 'living-room *objects* *object-locations*)
(YOU SEE A WHISKEY ON THE FLOOR. YOU SEE A BUCKET ON THE FLOOR.)