Added a "look" function to allow the user get a description of where he/she is.
This new function ties all together for the game. It's capable of describing
any location where the user is (determined by *location*). The game is ready to
roll

[1]> (look)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS
A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU
SEE A WHISKEY ON THE FLOOR. YOU SEE A BUCKET ON THE FLOOR.)