Added "walk" function.
All the pieces are fitting together.

[1]> (look)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU SEE A WHISKEY ON THE
 FLOOR. YOU SEE A BUCKET ON THE FLOOR.)
[2]> (walk 'west)
(YOU ARE IN A BEATIFUL GARDEN. THERE IS A WELL IN FRONT OF YOU THERE IS A DOOR GOING EAST FROM HERE. YOU SEE A CHAIN ON THE FLOOR.)
[3]> (walk 'east)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU SEE A WHISKEY ON THE
 FLOOR. YOU SEE A BUCKET ON THE FLOOR.)
[4]> (walk 'upstairs)
(YOU ARE IN THE ATTIC. THERE IS A GIGANT WELDING TORCH IN THE CORNER. THERE IS A LADDER GOING DOWNSTAIRS FROM HERE.)
[5]> (walk 'downstairs)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU SEE A WHISKEY ON THE
 FLOOR. YOU SEE A BUCKET ON THE FLOOR.)
[6]>

However, writing this last function was not that easy to understand, due to the
amount of parenthesis. I guess that in the end I will eat parenthesis for
breakfast, and eat more parenthesis for lunch!