Added pickup function, so one can take an object from the scenario.
Is interesting how the combinaiton of assoc and push behaves. In the new
function, to represent that the player has picked an object, we add at the
begining of the alist *object-locations* the choosen element and a new location
'body. The story, is that, when we call assoc to see where an object is, assoc,
returns the first entry it enccounters. So, despite, in this example I have in
the *object-location* list (after calling pickup), the object whiskey twice,
when querying where this object is, assoc will say 'body instead of saying
'body and 'living-room.

[1]> (look)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU SEE A WHISKEY ON THE
 FLOOR. YOU SEE A BUCKET ON THE FLOOR.)
[2]> (pickup 'whiskey)
(YOU ARE NOW CARRIYING THE WHISKEY)
[3]> (look)
(YOU ARE IN THE LIVING-ROOM. A WIZZARD IS SNORING LOUDLY ON THE COUCH. THERE IS A DOOR GOING WEST FROM HERE. THERE IS A LADDER GOING UPSTAIRS FROM HERE. YOU SEE A BUCKET ON THE
 FLOOR.)