Changed read for read-line
When using the regular "read" function, one need to specify the type of the
input. So, if in the example you wanted to introduce your name, and wanted it
to be treated as a string (as anyone would expect from a greeter), then you
need to specify is a string by adding double quotes when doing the input.

[1]> (say-hello)
Please type your name: Arturo
Nice to meet you, ARTURO
ARTURO
[2]> (say-hello)
Please type your name: "Arturo"
Nice to meet you, Arturo
"Arturo"
[3]>

Which is a bit of a hassle for the user. I have changed the read function for a
read-line function, so that it always expects a string as an input.

Print and Read functions in lisp has an interesting properties though, and it's
recommended to start with them before evolving the program to more targeted
options like princ, prin1, read-line, etc.