Improving read to accept plain commands.
First step to make an easy to use interface for the game is to let the user put
plain commands forgeting about the parenthesis. He should not be aware he's
calling a lisp function.
Thus, I have modified the read-game function, with some tips, from the book,
now, it takes a commands, and it adds the "(" and ")", so that it is
interpreted correctly.
It is still a bit ugly though...