hardening the custom game-read function
Now it forces to interpret as a symbol every extra parameter passed to the
commands. In this manner we avoid the user to input arbitrary functions calls
in the parameters. I.E, if the user inputs the following bad intentioned
commands:
   walk (format-disk)

Read will transform this into:
   (walk '(format-disk))

So that when eval is called, it does not evaluate the parameter as data
'(format-disk) instead of calling the function format-disk.