Printing embelishment :-)
After adding a couple of functions, the player is able to see the output from
the program as propper lowercases. In adition to this, the player does not need
to know which are objects and which are symbols, he just enters texts. Example
from a run:

[1]> (game-repl)
look
You are in the living-room. A wizzard is snoring loudly on the couch. There is a door going west from here. There is a ladder going upstairs from here. You see a whiskey on the
 floor. You see a bucket on the floor.
walk upstairs
You are in the attic. There is a gigant welding torch in the corner. There is a ladder going downstairs from here.
items
I do not know that command.
pickup torch
You cannot get that
pickup gigant
You cannot get that
walk downstairs
You are in the living-room. A wizzard is snoring loudly on the couch. There is a door going west from here. There is a ladder going upstairs from here. You see a whiskey on the
 floor. You see a bucket on the floor.
pickup whiskey
You are now carriying the whiskey
walk west
You are in a beatiful garden. There is a well in front of you there is a door going east from here. You see a chain on the floor.
pickup chain
You are now carriying the chain
walk east
You are in the living-room. A wizzard is snoring loudly on the couch. There is a door going west from here. There is a ladder going upstairs from here. You see a bucket on the
 floor.
inventory
Items- whiskey chain

Basically what I've done is to convert what needs to be printed into a string,
remove its parenthesis, and transform that into a list of characters. That is
passed to a "list eater" called tweak-text, that is putting the text in upper
or lower case. Last, the output from tweak-text is coerced back into string,
and passed to printc, which is printing it in the console.