Add function dot->png, that generates a png.
The purpse of this function is to write a digraph spec (in dot format) into a
file, and then call dot to create a png.
The function is designed to be generic, so it expects that the caller provides
a thunk, where the action of generating the dot from something specific is
passed. The idea is to leave this function just to do the IO part, all the
other logic, is wrapped in the thunk.