Added graph->png function, this function is called with a filename, an alist with the nodes, and alist with the edges.
This function is rather simple, and is the one which implements the thunk in a
lamda function, which is simply calling to graph->dot with the provided alists
for nodes and edges.
It has a little problem though, and is about the filename, the filename we
request is the dot filename, from which it generates a png using the file name
plus the extension .png. Which represents a minor problem, for instance if the
user wants to creat "test.png", the function will create a "test.png" with the
dot graph (in text) and a file called "test.png.png" to contain the actual png.