Attempt to avoid the filenaming problem when using the graph->png function
graph->png used the "fname" parameter to pass it as it is to dot->png, which
used that file name to print a digraph, and use it to call "dot", which in turn
will add a postfix (.png) to the filename.
Thus when we called
   (graph->png "test.png" *nodes* *edges*)

After its execution, a file test.png and a file test.png.png will be generated.
The problem is, that the user could try to open test.png and find out that it
does not work (since it is not a png, is a text file with the dot
description!).

To avoid the problem, I have changed the name of the parameter from fname to
prefix-fname, and I have also made that in graph->dot adds a ".dot" extension
to the file.