New approach to the problem, this time if the file have a .png extension, I replace it with .dot when generating the .dot file
With this new approach, two files are generated a file with the extension .dot
and a file with the extension .png, but we will remove the possibility of
having .png.png and .png.dot.

This is a result of an execution.

[1]> (graph->png "prueba.png" *nodes* *edges* )
T
[2]> (quit)
Bye.
darkturo@neverland:~/Learning/the-land-of-lisp/drawing-graphs$ ls
digraphs.lisp  prueba.dot     prueba.dot.png

For any other name the user may use, a .dot extension will be added to the
provided name when genrating the graph. And the same will happen when
generating the png, a ".png" extension will be used, i.e.

> (graph->png "superdoc.doc" *nodes* edges*)
> (ext:shell "ls")
superdoc.doc.dot superdoc.doc.png