Fixed little issue when calling graph->png, it returned nil.
(ext:shell) returns the return status of the executed command. In unix, this
status is by convention 0 when the command was executed correcly, or a
different number when it had failed for some reason. This behavior has been
passed to ext:shell, so it returns nil when it was okay, or a number when not.
Since we are encapsulating what we're doing inside the func,, I preffered to
encapsulate this detail and instead of showing nil or numbers, show true when
the exec was ok, and false (nil) when it didnt.