Fixed error in the generated dot file
ugraph->dot had an error, that made graphviz to fail while generating the png
file.

   [2]> (ugraph->png "prueba.dot" *wizzard-nodes* *wizzard-edges* )
   Error: prueba.dot.dot:5: syntax error near line 5
   context:  >>> GARDEN- <<< -LIVING_ROOM[label="(EAST DOOR)"];
   NIL

The problem was that I was trying to generate a digraph and setting links in the form "--".

   digraph {
   LIVING_ROOM[label="(LIVING-ROOM (YOU ARE IN TH..."];
   GARDEN[label="(GARDEN (YOU ARE IN A BEATI..."];
   ATTIC[label="(ATTIC (YOU ARE IN THE ATTI..."];
   GARDEN -- LIVING_ROOM[label="(EAST DOOR)"];
   ATTIC -- LIVING_ROOM[label="(DOWNSTAIRS LADDER)"];}

In graphviz the digraph keyword sets a context, where the links have to be
defined with arrows "->", but not with "--" straight lines. I did a mistake
while basing this function in the previous graph->dot function...
Now the problem is fixed, instead of using the keyword digraph, I use the
keyword "graph", which sets the right context to draw undirected graphs.