Extension of the "ext removal" feat.
The functions graps->png and ugraph->png in the original book example, were
expecting a file with extension ".dot", so that they will generate that file,
pass it to graphviz (dot) and let graphviz add a extension .png to the original
file. With that behavior if the user introduced "prueba.dot", after calling the
function, a file called "prueba.dot" and a file called "prueba.dot.png" were
created.

With this a extrange situation could happen if the user introduced
"prueba.png". Then the files "prueba.png" and "prueba.png.png" would have been
created. And that could be confusing to the user who may try to open prueba.png
and find it doesn't work.

So the idea was to find when the user introduced a "png" extension and try to
remove it. In the same way, I was adding a ".dot" extension to the file
generated with the graph. Some kind of normalization.
However, as a result, when calling the functions, the following files were created

   | Input         | dot file name generated | png file name generated |
   | ------------- | ----------------------- | ----------------------- |
   | "prueba.dot"  | prueba.dot.dot          | prueba.dot.dot.png      |
   | "prueba.png"  | prueba.dot              | prueba.dot.png          |
   | "prueba"      | prueba.dot              | prueba.dot.png          |
   | "prueba.doc"  | prueba.doc.dot          | prueba.doc.dot.png      |

With this commit, I'm changing the behavior so that the program does not
generate more extensions than the necessary, the behavior has been changed now
to do the following:

   | Input         | dot file name generated | png file name generated |
   | ------------- | ----------------------- | ----------------------- |
   | "prueba.dot"  | prueba.dot              | prueba.png              |
   | "prueba.png"  | prueba.dot              | prueba.png              |
   | "prueba"      | prueba.dot              | prueba.png              |
   | "prueba.doc"  | prueba.doc.dot          | prueba.doc.png          |

A sample of the execution (adding an echo at the begining of the cmd):

   [2]> (graph->png "filpando.png" *wizzard-nodes* *wizzard-edges*)
   dot -T png -o filpando.png filpando.dot
   T
   Break 1 [2]> (graph->png "filpando.dot" *wizzard-nodes* *wizzard-edges*)
   dot -T png -o filpando.png filpando.dot
   T
   Break 1 [2]> (graph->png "filpando" *wizzard-nodes* *wizzard-edges*)
   dot -T png -o filpando.png filpando.dot
   T
   Break 1 [2]> (graph->png "filpando.doc" *wizzard-nodes* *wizzard-edges*)
   dot -T png -o filpando.doc.png filpando.doc.dot