Added a bunch of functions to find and link unconnected nodes
The topology of the graph used as a city, is traced randomly. With this
technique the possibility of having unnconected nodes is quite hihg, so
with the functions created, the idea is to track these unconnected nodes
or islands, and trace edges (or links) with at least one connected node.

The main function, here is connect-all-islands, which uses all the
auxiliary functions connect-with-bridges, find-islands, get-connected,
and direct-edges.